+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Make `/api/v1/apps` and `/oauth/token` endpoints accept `multipart/form-data`.
 - Don't retry incoming activity if fetcher encounters `404 Not Found` error.
 - Rename `check-expired-invoice` command to `reopen-invoice` and allow to reopen invoices with "forwarded" status.
+- Allow to call `reopen-invoice` command with payment address as an argument.
 
 ### Deprecated